+++ lib/parse/model/core/actions.rb
@@ -501,6 +501,10 @@ module Parse
 
       # saves the object. If the object has not changed, it is a noop. If it is new,
       # we will create the object. If the object has an id, we will update the record.
+      #
+      # You may pass a session token to the `session` argument to perform this actions
+      # with the priviledges of a certain user.
+      #
       # You can define before and after :save callbacks
       # autoraise: set to true will automatically raise an exception if the save fails
       # @raise Parse::RecordNotSaved if the save fails